DAN AI/requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
requests==2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
